feat(Filters): fix sonar
